-Created Activity service -Added get a user funciton to the user service

+++ CorporateQnA.Services/Models/AnswerActivity.cs
@@ -22,7 +22,7 @@ namespace CorporateQnA.Services.Models
         public int AnswerId { get; set; }
 
         [Column]
-        public int ActivityType { get; set; }
+        public short ActivityType { get; set; }
 
         [Column]
         public DateTime CreatedAt { get; set; }
